Removed hint* methods from Graph API
Instead, the size_hint() method of the corresponding iterator should be used.
This is more idiomatic, and eases implementation since complex iterators (e.g. filter, map) usually provide good implementations of size_hint.
This adds a slight overhead (since the iterators of the graph API are boxed trait objects), but this is probably insignificant in the situations where a size-hint is required.