simplified TermMatcher
Removed the type parameter in TermMatcher's definition,
which was really only needed for the 'Fn(Term<T>) -> bool' implementation.

Instead, I limited the latter by 'Fn(RefTerm) -> bool', which is ok because
any term can be converted to a RefTerm.

The benefit is that TermMatcher with no type parameter will be easier to use
in the future refactoring.