Graph now has a lifetime parameter
This is a first step towards a more generic trait,
where Graph iterators will be able to yield any implementation of Triple.

This makes the relation with MutableGraph a little more complex.
A consequence is that MutableGraph can only be implemented
if Graph is implemented for *any* lifetime
(in short: a Graph may borrow its data, but a MutableGraph must own them).
Another consequence is that MutableGraph now has its own associated type
for errors.

The latter point forced me to ignore some errors in remove_matching and retain;
this should be addressed later, probably using the chaining mechanisms of
error_chain of failure (if I decide to migrate).