Graph now has an associated type Triple
This is more generic than the previous version,
where the triples produced by a Graph were necessarily tuples of &Term.

This required to refactor the IndexMutableGraph trait,
and the types in the graph::inmem modules.

Finally, I also deprecated the use of (Term,Term,Term) as Triple implementation,
in favor of [Term;3].