Now using coercible_errors
The goal is to avoid overhead for infallible implementations
of the various traits defined in sophia (Graph, MutableGraph, TripleSink,
etc).

Note that the API of Serializers has slightly changed in the process:
stringifying may now fail (because not all syntaxes support all graphs,
especially non-standard graphs -- e.g. literals as subjects).

Note also that MutableGraph::remove_matching and MutableGraph::retain
are not yet optimal; they may panic if the underyling iteration fails.
This should be fixed in the future -- I'm hoping that non-lexical
lifetimes may help (after the migration to the 2018 edition).