improved MutableGraph::remove_all and MutableGraph::retain
The problem was *not* lifetimes after all.
The trick was to use
        for <'a> <Self as Graph<'a>>::Error: Into<Self::MutationError>,
instead of
        for <'a> Self::MutationError: From<<Self as Graph<'a>>::Error>,
(the latter being not inferred by the compiler, for some reason...
The drawback is that I had to resort to an explicit
	.map_err(|e| e.into())
before lifting the error with '?'.
But at least all errors are correctly handled now.