Added methods to the Graph trait
These methods are subject, predicates, objects, iris, bnodes, literals
and variables. They return HashSets of terms. The default implementations
build those terms by naively iterating over all triples of the graph.

Because the Graph trait is now so generic, we can not assume that these
terms will exist as long as the graph does (terms in triples may be
generated on demand when the triple is required).
So these methods return a HashSet<Term> rather than a HashSet<&Term>,
forcing to *clone* the terms from the Graph.
In most cases, this is not too bad because Terms themselves contain
smart pointers (Rc<str>, Arc<str>) and so cloning them is still relatively
cheap.