Graph and Dataset no longer have a liferime paramerer
This is achieved by introducing the notion of "streming mode"
for triples and quads, and Graph::Triple (resp. Dataset::Quad)
are no longer the direct type of Triple/Quad return by their methods,
but the streaming mode used to determine the actual type.

Once Generic Assiociated Types are stabilized,
this notion of "streaming mode" will no longer be required,
and will be deprecated. But I don't expect this to happen soon.