Simplified Parser API
I removed the generic Parser trait.
TripleParser and QuadParser are now independent (although similar) traits.

I did it because Parser didn't make much sense from the users' point of
view, and yet it is the trait that they had to implement, in order to
"magically" get the TripleParser or QuadParser implemented.
Now they have to directly implement TripleParser or QuadParser, which I
think is much clearer and straightforward.

The main reason for having a generic trait in the first place was to
reduce code duplication, but there is already a lot of code duplication
between triple- and quad-related code, and the gain here was tiny
(although we might want to add other methods in the future...).