add MownStr type (maybe owned string) and MownTerm
Just like RefTerm, a MownTerm can be constructed from any Term,
simply borrowing the original term's data.
But a MownTerm can also own its data (or part of it).

A use-case for this is Resolve;
when resolving a collection of terms to MownTerms,
the source terms that were already absolute need not be copied,
while those that were relative will need some allocation
for their resolved counterpart.
Note that this is not implemented yet
(currently, Resolve cannot produce MownTerms,
because MownStr does not implement `for <'a> From<&'a str>`).