WIP musing with the idea of a trait object
I like the fact that many types (including atomic ones) could then be used as terms.
For that reason, I decided to make MownStr the rule, so that some types may return references,
while others may build and return owned values on demand (example: the value of i32 as Literal).