add feature all_tests
A number of tests are now hidden behind this feature, which not enabled by
default. This make the compilation of tests significantly faster.

The tests that are skipped are mostly redundant with other tests, which
makes it ok to skip in development iterations.

Obviously, 'cargo test --all-features' should still be ran before pushing
any commit...