improve performance of Rio adapters
instead of converting Rio terms to sophia_term::Term,
we now provide a simple adapter for the TTerm trait,
which significantly increses performances (~10% and ~20%).

NB: we do not rely on the `sophia` feature of Rio,
which does the same,
but using the previously published version of Sophia.
If we want `sophia_api` and `sophia` to be published in sync,
and `sophia` to depend on Rio,
we can not depend on these adapters.