upgrade spargebra to 0.4.0-beta.1
this one is based on RDF 1.2 rather than RDF-star,
so this should probably be done once the rest of Sophia is adapted to RDF 1.2
(i.e. supports language strings, and possibly has RDF 1.2 parsers/serializers)

note that the changes in sparql/src/test.rs and sparql/src/function/test.rs
were entirely automated: replace '<< ' with '<<( ' and ' >>' with ' )>>'.