refactor some foreign impl of Graph and Dataset
... to now return references to their internal Triples/Quads.
This may break some implementation that relied on the fact that
these impls yielded [Term;3] or Spog<Term>, respectively.
A simple fix is to call Triple::spo or Quad::spog
(or possibly Triple::to_spo or Quad::to_spog)
as can be seen in turtle/src/serializer/_pretty.rs:440 in this commit.