fix scoping issue in SPARQL evaluation
SPARQL evaluation is defined bottom-up.
In particular, for a join, both sides are evaluated,
and only the compatible mappings are combined and returned.

However, for optimization,
this implementation is designed to "push down" the left binding ("context"),
to constrain the evaluation of the right binding.

The way this optimization was implemented broke SPARQL semantics,
when expressions are involved:
the variables bound in the context should *not* participate in the evaluation of the expression.

So, for all algebraic construct involving an expression,
we strictly apply the bottom-up evaluation logic;
the context is not even passed anymore to the corresponding functions.

For other algebraic constructs, we continue pushing down the context,
but it is now clearly document that this is just a hint,
and that the returned bindings are not guaranteed to be compatible or merged with the context.
It will be the responsibility of join and left_join to actually do it.